feat: change path from guildName to guildSlug
