+++ app/g/[guildSlug]/layout.tsx
@@ -1,0 +1,22 @@
+'use client';
+
+import GuildInfo from '@/components/guilds/guild-info';
+
+export default function GuildLayout({
+  children,
+  params,
+}: {
+  children: React.ReactNode;
+  params: {
+    guildSlug: string;
+  };
+}) {
+  return (
+    <div className="max-w-5xl mx-auto">
+      <div className="pt-4 pb-2">
+        <GuildInfo slug={params.guildSlug} />
+      </div>
+      {children}
+    </div>
+  );
+}
