+++ app/g/[guildSlug]/market/demands/[categorySlug]/layout.tsx
@@ -9,7 +9,7 @@ interface Props extends DemandsPageProps {
 function DemandsLayout({ params, children }: Props) {
   return (
     <div className="pb-4 md:pb-6">
-      <CategoriesNavbar guildName={params.guildName} />
+      <CategoriesNavbar categorySlug={params.categorySlug} />
       <div className="grid gap-1 grid-cols-1 mx-2 md:mx-0">{children}</div>
     </div>
   );
