+++ app/g/[guildSlug]/market/swaps/[categorySlug]/layout.tsx
@@ -9,7 +9,7 @@ interface Props extends SwapsPageProps {
 function SwapsLayout({ params, children }: Props) {
   return (
     <div className="pb-4 md:pb-6">
-      <CategoriesNavbar guildName={params.guildName} />
+      <CategoriesNavbar categorySlug={params.categorySlug} />
       <div className="grid gap-x-0 md:gap-x-6 gap-y-1 lg:gap-y-14 grid-cols-1 md:grid-cols-3 lg:grid-cols-3 mx-2 md:mx-0">
         {children}
       </div>
